river_collector/requirements.txt
requests>=2.32.3,<2.33
geopandas>=1.0.1,<1.1
python-dotenv>=1.0.1,<1.1
geopy>=2.4.1,<2.5
folium>=0.18.0,<0.19
